Fix CI/CD: Add size_t casts in Labellist operator[] methods
Fixed sign conversion warnings in lparser.h operator[] overloads:
- Cast index parameter to size_t when accessing vec[index]
- Applied to both mutable and const versions of operator[]

This resolves sign conversion warnings from int to unsigned long.

All struct/class mismatches and sign conversions should now be fixed!

+++ src/compiler/lparser.h
@@ -231,8 +231,8 @@ public:
   /* Direct vector access for modern operations */
   inline void push_back(const Labeldesc& desc) { vec.push_back(desc); }
   inline void reserve(int capacity) { vec.reserve(static_cast<size_t>(capacity)); }
-  inline Labeldesc& operator[](int index) { return vec[index]; }
-  inline const Labeldesc& operator[](int index) const { return vec[index]; }
+  inline Labeldesc& operator[](int index) { return vec[static_cast<size_t>(index)]; }
+  inline const Labeldesc& operator[](int index) const { return vec[static_cast<size_t>(index)]; }
 
   /* For luaM_growvector replacement */
   inline void ensureCapacity(int needed) {
